Rename file for chart comparison options
Rename file for chart comparison options, from ChartComparisonHelpers to chartComparisonOptions to avoid conflict with ChartComparisonHelpers.
