+++ client/src/helper/chartComparisonOptions.js
